requirements.txt
absl-py==1.0.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
Django==3.2.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
dlib==19.23.1
entrypoints==0.4
executing==0.8.3
flatbuffers==2.0
fonttools==4.28.5
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.45.0
h5py==3.6.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.11.3
imutils==0.5.4
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
networkx==2.8
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.0.0
pip==21.2.4
prompt-toolkit==3.0.27
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
setuptools==58.0.4
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.4.8
toml==0.10.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.14.0
zipp==3.7.0
